Pull from upstream.

+++ Library/Formula/qt.rb
@@ -1,84 +1,108 @@
 require 'formula'
-require 'hardware'
 
-class Qt <Formula
-  url 'http://get.qt.nokia.com/qt/source/qt-everywhere-opensource-src-4.7.1.tar.gz'
-  md5 '6f88d96507c84e9fea5bf3a71ebeb6d7'
-  homepage 'http://qt.nokia.com/'
+class Qt < Formula
+  homepage 'http://qt-project.org/'
+  url 'http://releases.qt-project.org/qt4/source/qt-everywhere-opensource-src-4.8.4.tar.gz'
+  sha1 'f5880f11c139d7d8d01ecb8d874535f7d9553198'
+
+  bottle do
+    revision 1
+    sha1 '7fb679119b8b463055849dea791cc7fca62c62d1' => :mountain_lion
+    sha1 'b456ff5f8d18fc53b4546119d00d8ff0dda92f90' => :lion
+    sha1 '920992e5059a5c816b4eb245597fc028ff6b09ae' => :snow_leopard
+  end
+
+  head 'git://gitorious.org/qt/qt.git', :branch => 'master'
+
+  option :universal
+  option 'with-qtdbus', 'Enable QtDBus module'
+  option 'with-qt3support', 'Enable deprecated Qt3Support module'
+  option 'with-demos-examples', 'Enable Qt demos and examples'
+  option 'with-debug-and-release', 'Compile Qt in debug and release mode'
+  option 'developer', 'Compile and link Qt with developer options'
+
+  depends_on :libpng
+
+  depends_on "d-bus" if build.with? 'qtdbus'
+  depends_on "mysql" => :optional
+  depends_on 'sqlite' if MacOS.version == :leopard
 
   def patches
-    # To fix http://bugreports.qt.nokia.com/browse/QTBUG-13623. Patch sent upstream.
-    "http://qt.gitorious.org/qt/qt/commit/9f18a1ad5ce32dd397642a4c03fa1fcb21fb9456.patch"
+    # Fixes compilation failure on Leopard.
+    # https://bugreports.qt-project.org/browse/QTBUG-23258
+    if MacOS.version == :leopard
+      "http://bugreports.qt-project.org/secure/attachment/26712/Patch-Qt-4.8-for-10.5"
+    end
   end
-
-  def options
-    [
-      ['--with-qtdbus', "Enable QtDBus module."],
-      ['--with-qt3support', "Enable deprecated Qt3Support module."],
-      ['--with-demos-examples', "Enable Qt demos and examples."],
-      ['--with-debug-and-release', "Compile Qt in debug and release mode."],
-      ['--universal', "Build both x86_64 and x86 architectures."],
-    ]
-  end
-
-  def self.x11?
-    File.exist? "/usr/X11R6/lib"
-  end
-
-  depends_on "d-bus" if ARGV.include? '--with-qtdbus'
-  depends_on 'libpng' unless x11?
-  depends_on 'sqlite' if MACOS_VERSION <= 10.5
 
   def install
     ENV.append "CXXFLAGS", "-fvisibility=hidden"
+
+    # clang complains about extra qualifier since Xcode 4.6 (clang build 425)
+    # https://bugreports.qt-project.org/browse/QTBUG-29373
+    if MacOS.clang_build_version >= 425
+      inreplace "src/gui/kernel/qt_cocoa_helpers_mac_p.h",
+                "::TabletProximityRec",
+                "TabletProximityRec"
+    end
+
     args = ["-prefix", prefix,
             "-system-libpng", "-system-zlib",
             "-confirm-license", "-opensource",
             "-cocoa", "-fast" ]
 
-    # See: https://github.com/mxcl/homebrew/issues/issue/744
-    args << "-system-sqlite" if MACOS_VERSION <= 10.5
-    args << "-plugin-sql-mysql" if (HOMEBREW_CELLAR+"mysql").directory?
+    # we have to disable 3DNow! to avoid triggering optimization code
+    # that will fail with clang. Only seems to occur in superenv, perhaps
+    # because we rename clang to cc and Qt thinks it can build with special
+    # assembler commands. In --env=std, Qt seems aware of this.)
+    # But we want superenv, because it allows to build Qt in non-standard
+    # locations and with Xcode-only.
+    args << "-no-3dnow" if superenv?
 
-    if ARGV.include? '--with-qtdbus'
+    args << "-L#{MacOS::X11.prefix}/lib" << "-I#{MacOS::X11.prefix}/include" if MacOS::X11.installed?
+
+    args << "-platform" << "unsupported/macx-clang" if ENV.compiler == :clang
+
+    # See: https://github.com/mxcl/homebrew/issues/issue/744
+    args << "-system-sqlite" if MacOS.version == :leopard
+
+    args << "-plugin-sql-mysql" if build.with? 'mysql'
+
+    if build.with? 'qtdbus'
       args << "-I#{Formula.factory('d-bus').lib}/dbus-1.0/include"
       args << "-I#{Formula.factory('d-bus').include}/dbus-1.0"
       args << "-L#{Formula.factory('d-bus').lib}"
       args << "-ldbus-1"
-      args << "-dbus-linked"
     end
 
-    if ARGV.include? '--with-qt3support'
+    if build.with? 'qt3support'
       args << "-qt3support"
     else
       args << "-no-qt3support"
     end
 
-    if ARGV.include? '--with-debug-and-release'
+    unless build.with? 'demos-examples'
+      args << "-nomake" << "demos" << "-nomake" << "examples"
+    end
+
+    if MacOS.prefer_64_bit? or build.universal?
+      args << '-arch' << 'x86_64'
+    end
+
+    if !MacOS.prefer_64_bit? or build.universal?
+      args << '-arch' << 'x86'
+    end
+
+    if build.with? 'debug-and-release'
       args << "-debug-and-release"
+      # Debug symbols need to find the source so build in the prefix
+      mv "../qt-everywhere-opensource-src-#{version}", "#{prefix}/src"
+      cd "#{prefix}/src"
     else
       args << "-release"
     end
 
-    unless ARGV.include? '--with-demos-examples'
-      args << "-nomake" << "demos" << "-nomake" << "examples"
-    end
-
-    if Qt.x11?
-      args << "-L/usr/X11R6/lib"
-      args << "-I/usr/X11R6/include"
-    else
-      args << "-L#{Formula.factory('libpng').lib}"
-      args << "-I#{Formula.factory('libpng').include}"
-    end
-
-    if snow_leopard_64? or ARGV.include? '--universal'
-      args << '-arch' << 'x86_64'
-    end
-
-    if !snow_leopard_64? or ARGV.include? '--universal'
-      args << '-arch' << 'x86'
-    end
+    args << '-developer-build' if build.include? 'developer'
 
     system "./configure", *args
     system "make"
@@ -92,17 +116,32 @@ class Qt <Formula
     (bin+'pixeltool.app').rmtree
     (bin+'qhelpconverter.app').rmtree
     # remove porting file for non-humans
-    (prefix+'q3porting.xml').unlink
+    (prefix+'q3porting.xml').unlink if build.without? 'qt3support'
 
     # Some config scripts will only find Qt in a "Frameworks" folder
-    # VirtualBox is an example of where this is needed
-    # See: https://github.com/mxcl/homebrew/issues/issue/745
-    cd prefix do
-      ln_s lib, "Frameworks"
+    frameworks.mkpath
+    ln_s Dir['lib/*.framework'], frameworks
+
+    # The pkg-config files installed suggest that headers can be found in the
+    # `include` directory. Make this so by creating symlinks from `include` to
+    # the Frameworks' Headers folders.
+    Pathname.glob(lib + '*.framework/Headers').each do |path|
+      framework_name = File.basename(File.dirname(path), '.framework')
+      ln_s path.realpath, include+framework_name
+    end
+
+    Pathname.glob(bin + '*.app').each do |path|
+      mv path, prefix
     end
   end
 
-  def caveats
-    "We agreed to the Qt opensource license for you.\nIf this is unacceptable you should uninstall."
+  test do
+    system "#{bin}/qmake", '-project'
+  end
+
+  def caveats; <<-EOS.undent
+    We agreed to the Qt opensource license for you.
+    If this is unacceptable you should uninstall.
+    EOS
   end
 end
